nvapi: Move Vk to shared
Also split current ResourceFactory into ResourceFactory and
NvapiResourceFactory to keep Shared free from dxgi. lfx and
nvml.
