+++ src/meson.build
@@ -2,18 +2,19 @@ nvapi_src = files([
   'util/util_string.cpp',
   'util/util_env.cpp',
   'util/util_log.cpp',
-  'nvapi/resource_factory.cpp',
+  'shared/vk.cpp',
+  'shared/resource_factory.cpp',
   'nvapi/lfx.cpp',
+  'nvapi/nvml.cpp',
+  'nvapi/nvapi_resource_factory.cpp',
   'nvapi/nvapi_d3d_instance.cpp',
   'nvapi/nvapi_d3d_low_latency_device.cpp',
   'nvapi/nvapi_vulkan_low_latency_device.cpp',
+  'nvapi/nvapi_output.cpp',
+  'nvapi/nvapi_adapter.cpp',
+  'nvapi/nvapi_adapter_registry.cpp',
   'nvapi/nvapi_d3d11_device.cpp',
   'nvapi/nvapi_d3d12_device.cpp',
-  'nvapi/nvapi_adapter_registry.cpp',
-  'nvapi/nvapi_adapter.cpp',
-  'nvapi/nvapi_output.cpp',
-  'nvapi/vk.cpp',
-  'nvapi/nvml.cpp',
   'nvapi_globals.cpp',
   'nvapi.cpp',
   'nvapi_d3d.cpp',
