+++ src/nvapi/nvapi_adapter_registry.cpp
@@ -7,7 +7,7 @@
 
 namespace dxvk {
 
-    NvapiAdapterRegistry::NvapiAdapterRegistry(ResourceFactory& resourceFactory)
+    NvapiAdapterRegistry::NvapiAdapterRegistry(NvapiResourceFactory& resourceFactory)
         : m_resourceFactory(resourceFactory) {}
 
     NvapiAdapterRegistry::~NvapiAdapterRegistry() {
