+++ src/nvapi/nvapi_resource_factory.h
@@ -2,20 +2,18 @@
 
 #include "../nvapi_private.h"
 #include "../util/com_pointer.h"
-#include "vk.h"
+#include "../shared/resource_factory.h"
+#include "../shared/vk.h"
 #include "nvml.h"
 #include "lfx.h"
 
 namespace dxvk {
-    class ResourceFactory {
+    class NvapiResourceFactory : ResourceFactory {
 
       public:
-        ResourceFactory();
-        virtual ~ResourceFactory();
-
-        virtual Com<IDXGIFactory1> CreateDXGIFactory1();
+        std::unique_ptr<Vk> CreateVulkan(const char* moduleName) override { return ResourceFactory::CreateVulkan(moduleName); }
         virtual std::unique_ptr<Vk> CreateVulkan(Com<IDXGIFactory1>& dxgiFactory);
-        virtual std::unique_ptr<Vk> CreateVulkan(const char* moduleName);
+        virtual Com<IDXGIFactory1> CreateDXGIFactory1();
         virtual std::unique_ptr<Nvml> CreateNvml();
         virtual std::unique_ptr<Lfx> CreateLfx();
     };
