+++ tests/meson.build
@@ -2,10 +2,11 @@ nvapi_src = files([
   '../src/util/util_string.cpp',
   '../src/util/util_env.cpp',
   '../src/util/util_log.cpp',
-  '../src/nvapi/resource_factory.cpp',
-  '../src/nvapi/vk.cpp',
+  '../src/shared/vk.cpp',
+  '../src/shared/resource_factory.cpp',
   '../src/nvapi/nvml.cpp',
   '../src/nvapi/lfx.cpp',
+  '../src/nvapi/nvapi_resource_factory.cpp',
   '../src/nvapi/nvapi_d3d_instance.cpp',
   '../src/nvapi/nvapi_d3d_low_latency_device.cpp',
   '../src/nvapi/nvapi_vulkan_low_latency_device.cpp',
