+++ tests/nvapi_vulkan_mocks.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "nvapi_tests_private.h"
-#include "../src/nvapi/vk.h"
+#include "../src/shared/vk.h"
 
 using namespace trompeloeil;
